Finished Readme Description
This is the final description for the mod until content is implemented and fixed, commits will probably begin to be added relatively soon. I dont have plans to translate any of this work or start in EU5 as I still much more enjoy EU4 as a game and with its potential for modding. Mods like Post Finem (shout out chewyshoot) show me that it is still worth making things for my favourite game.